refactor: separate result component from API function
