+++ src/main.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './app';
-import './index.scss';
+import App from './pages';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
