+++ src/service/types.tsx
@@ -16,3 +16,7 @@ export type State = {
   hasError: boolean;
   error: Error | null;
 };
+
+export type AppState = {
+  searchTerm: string;
+};
